+++ composeApp/src/commonMain/kotlin/com/github/singularity/core/shared/model/http/Response.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PairResponse(
     val success: Boolean,
-    val pairRequestId: Long? = null,
+    val pairRequestId: Int? = null,
 )
 
 @Serializable
